+++ src/ploomber/cli/build.py
@@ -1,3 +1,5 @@
+import sys
+
 from ploomber.cli.parsers import CustomParser
 from ploomber.cli.io import cli_endpoint
 from ploomber.executors import Parallel
@@ -9,7 +11,7 @@ import datetime
 @cli_endpoint
 def main(render_only=False):
     start_time = datetime.datetime.now()
-    parser = CustomParser(description='Build pipeline')
+    parser = CustomParser(description='Build pipeline', prog='ploomber build')
 
     with parser:
         parser.add_argument('--force',
@@ -34,6 +36,12 @@ def main(render_only=False):
             help='Drop a debugger session if an exception happens',
             action='store_true',
             default=False)
+
+    # users may try to run "ploomber build {name}" to build a single task
+    if len(sys.argv) > 1 and not sys.argv[1].startswith('-'):
+        suggestion = 'ploomber task {task-name}'
+        parser.error(f'{parser.prog!r} does not take positional arguments.\n'
+                     f'To build a single task, try: {suggestion!r}')
 
     dag, args = parser.load_from_entry_point_arg()