+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Fixes error message when failing to load dotted paths
 * `ploomber scaffold` now supports `.R and .Rmd` files (#476)
 * Fixes an error that caused `ploomber scaffold` to ignore the location of existing packages (#459)
+* Better error message when running `ploomber execute/run` (suggests `ploomber build`)
+* Better error message when passing positional arguments to `ploomber build` (suggests `ploomber task`)
 
 ## 0.14.4 (2022-01-07)
 * Fixes an error in the telemetry module